QMdmmBot: make the seat count narrowing explicit
players().count() returns qsizetype, so assigning it to the int used by the
canMove() loop narrowed implicitly. Cast it explicitly, matching the
static_cast<int> style used elsewhere in the tree. Room seat counts are tiny,
so int is the right type here.

Both styles do the same thing; found by clang-tidy once the missing -isysroot
was supplied.

+++ QMdmmBot/src/knifepreferredbot.cpp
@@ -54,7 +54,7 @@ void KnifePreferredBot::handleActionRequest(int currentOrder)
             return;
         }
         // Cannot buy where we stand (e.g. Village) -> step to any city seat.
-        const int seatCount = client()->room()->players().count();
+        const int seatCount = static_cast<int>(client()->room()->players().count());
         for (int place = 1; place <= seatCount; ++place) {
             if (self->canMove(place)) {
                 client()->agent()->action(QMdmmCore::Data::Move, QString(), place);
